backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
motor==3.6.0
pymongo==4.9.2
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.27.2
PyJWT==2.9.0
cryptography==43.0.1
mangum==0.17.0
python-multipart==0.0.9

